Break out reorder and resize of spreadsheet
